seeding

+++ database/migrations/2025_03_08_033458_add_asal_sekolah_to_ppdbs_table.php
@@ -10,17 +10,21 @@ return new class extends Migration
      * Run the migrations.
      */
     public function up()
-{
-    Schema::table('ppdbs', function (Blueprint $table) {
-        $table->string('asal_sekolah')->after('agama');
-    });
-}
+    {
+        if (!Schema::hasColumn('ppdbs', 'asal_sekolah')) {
+            Schema::table('ppdbs', function (Blueprint $table) {
+                $table->string('asal_sekolah')->after('agama');
+            });
+        }
+    }
 
-public function down()
-{
-    Schema::table('ppdbs', function (Blueprint $table) {
-        $table->dropColumn('asal_sekolah');
-    });
-}
+    public function down()
+    {
+        if (Schema::hasColumn('ppdbs', 'asal_sekolah')) {
+            Schema::table('ppdbs', function (Blueprint $table) {
+                $table->dropColumn('asal_sekolah');
+            });
+        }
+    }
 
 };
